AppArmor: move to 32 bit file permissions, allowing for future expansion
The 16 bit file permission space is exhausted, so expand to 32 bits
allowing for easy expansion in the future.

+++ include/linux/lsm_audit.h
@@ -114,8 +114,8 @@ struct common_audit_data {
 				} rlim;
 				struct {
 					const char *target;
-					u16 request;
-					u16 denied;
+					u32 request;
+					u32 denied;
 					uid_t ouid;
 				} fs;
 				struct {
